🐛 fix|Geometric|Clip triangles (walls) so they are rendered
Walls that are very tall or very deep in z elevation will not render when viewed on a slant, b/c they extend to the other side of the z plane. Clip so only the negative z remain, which causes the perspective matrix to correctly transform the values. Must clip after orthogonal but before perspective.